requirements.txt
alembic==1.13.2
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.1.0
gunicorn==23.0.0
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
SQLAlchemy==2.0.34
typing_extensions==4.12.2
Werkzeug==3.0.4
zipp==3.20.2
